collect all rejected names before throwing error
